% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Movie Explorer application}
\usage{
launch_app(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}
}
\value{
A Shiny app object (invisibly).
}
\description{
Entry point for the \code{movexplR6} package. Calls \code{\link[shiny:shinyApp]{shiny::shinyApp()}} with the
package UI and server functions.
}
\examples{
\dontrun{
launch_app()
}

}
\seealso{
\code{\link[=movies_ui]{movies_ui()}}, \code{\link[=movies_server]{movies_server()}}

Other Application Components:
\code{\link[=movies_server]{movies_server()}},
\code{\link[=movies_ui]{movies_ui()}}
}
\concept{Application Components}
