% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filters.R
\name{mod_filters_server}
\alias{mod_filters_server}
\title{Filter inputs module server}
\usage{
mod_filters_server(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A reactive list with elements: \code{reviews}, \code{oscars}, \code{year},
\code{boxoffice}, \code{genre}, \code{director}, \code{cast}, \code{xvar}, \code{yvar}.
}
\description{
Returns a reactive list of all current filter and axis selections.
}
