% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filters.R
\name{mod_filters_ui}
\alias{mod_filters_ui}
\title{Filter inputs module UI}
\usage{
mod_filters_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A \code{tagList} of \code{bslib::card} elements.
}
\description{
Renders two \code{bslib} cards: one for filter controls and one for axis
variable selectors. Intended for use inside a \code{bslib::sidebar()}.
}
