% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot.R
\name{mod_plot_ui}
\alias{mod_plot_ui}
\title{Scatter plot display module UI}
\usage{
mod_plot_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A \code{bslib::card}.
}
\description{
Renders a \code{bslib::card} containing a \code{plotly} scatter plot and a movie count.
}
