% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_server.R
\name{movies_server}
\alias{movies_server}
\title{Main server logic for the Movie Explorer application}
\usage{
movies_server(input, output, session)
}
\arguments{
\item{input, output, session}{Standard Shiny server arguments.}
}
\value{
No return value; called for side effects.
}
\description{
Initializes a \link{MovieData} R6 object, wires the filter module to the data
object's \verb{$filter()} method, and passes reactive results to the plot module.
}
\examples{
\dontrun{
shiny::shinyApp(ui = movies_ui(), server = movies_server)
}

}
\seealso{
\code{\link[=movies_ui]{movies_ui()}}, \code{\link[=launch_app]{launch_app()}}

Other Application Components:
\code{\link[=launch_app]{launch_app()}},
\code{\link[=movies_ui]{movies_ui()}}
}
\concept{Application Components}
